Implemented serialization from the bootcompiler to the bootloader.
Before, the bootcompiler was outputting .cc files which created
the big code areas in memory.

Now, it outputs .ozb which contain a pickled version of the same
code areas. Only stateless, structural data types can be contained
in this "boot" pickles. The engine can unpickle these rather
easily.

One good thing about this is that it is a first major step towards
a full support for serialization.

The other good thing is that, since we do not have ~40 .cc files
to compile anymore, the build process is *significantly* faster!