Added basic support for parse errors
Parse errors list the name of tokens or rules that were expected at the
point where they occurred. They also describe the state the parser was
in when it happened (for instance, if the error happened within a
local-statement, a method definition, or a record literal).

To achieve this, the grammar needs to be annotated with label(X Parser)
and context(X Parser).

This does have a very negative effect on parsing performance.