Parsing rules are now looked up at translate-time
The main purpose of this patch is to improve performance of the
parser. It does so significantly (and is in fact almost enough to make
the parser as fast as it was when it didn't handle errors).

Every time the parser encountered a non-terminal rule, it would look up
the procedure that needed to be called. Since this happens extremely
often, moving the code that does this to the time when the rules are
being translated (which only happens once) is very helpful.

Notice that this works even for mutually recursive rules when none of
them are yet defined thanks to unbound variables.

PEG.oz also allowed for each context to specify which grammar to
use. Since none of the current parsers used that feature, it has been
removed in the interest of simplicity and performance both (it could be
re-added with close to cost when it isn't being used with an option if
need be).